Fix translatePoints
Previously, translatePoints was returning an object with right and top.
Other methods, after calling translatePoints, were using left and top,
so those methods had to be overridden too. Now it returns an object
with left and top, but left is actually right. So, setXY of element is
overridden to handle this.